drivers: ipa_v3: Fix format specifier warning

+++ drivers/platform/msm/ipa/ipa_v3/ipa_qdss.c
@@ -137,8 +137,8 @@ int ipa3_conn_qdss_pipes(struct ipa_qdss_conn_in_params *in,
 	result = gsi_alloc_channel(&gsi_channel_props, ipa3_ctx->gsi_dev_hdl,
 				&ep_rx->gsi_chan_hdl);
 	if (result != GSI_STATUS_SUCCESS) {
-		IPA_QDSS_ERR("Failed allocating gsi_chan_hdl=%d\n",
-				&ep_rx->gsi_chan_hdl);
+		IPA_QDSS_ERR("Failed allocating gsi_chan_hdl=%ld\n",
+				(unsigned long)&ep_rx->gsi_chan_hdl);
 		goto fail;
 	}
